core/state/snapshot: fix review concerns
